Separate script utils
